renamed ticket.ts to leaderboard.ts and added code
