+++ server/src/models/index.ts
@@ -3,22 +3,22 @@ dotenv.config();
 
 import { Sequelize } from 'sequelize';
 import { UserFactory } from './user.js';
-import { TicketFactory } from './ticket.js';
+import { TicketFactory } from './leaderboard.js';
 
 const sequelize = process.env.DB_URL
   ? new Sequelize(process.env.DB_URL)
   : new Sequelize(process.env.DB_NAME || '', process.env.DB_USER || '', process.env.DB_PASSWORD, {
-      host: 'localhost',
-      dialect: 'postgres',
-      dialectOptions: {
-        decimalNumbers: true,
-      },
-    });
+    host: 'localhost',
+    dialect: 'postgres',
+    dialectOptions: {
+      decimalNumbers: true,
+    },
+  });
 
 const User = UserFactory(sequelize);
 const Ticket = TicketFactory(sequelize);
 
 User.hasMany(Ticket, { foreignKey: 'assignedUserId' });
-Ticket.belongsTo(User, { foreignKey: 'assignedUserId', as: 'assignedUser'});
+Ticket.belongsTo(User, { foreignKey: 'assignedUserId', as: 'assignedUser' });
 
 export { sequelize, User, Ticket };
